Consolidated .description and .forcegroups into single, unified .forces module
